RepairMode
Instead of use by [info host] in several places, I define ::repairMode=1 for DavidYashar, AvrahamBismut and so on.
If ::repairMode==1, you can define if Etx220 Generator exists or not. If not, the DataTransmission_conf and DataTransmission_run tests will be excluded from glTests.

To avoid "variable does not exist" error, I add to start of the GUI proc the following lines:
  if {![info exists ::repairMode]} {
    set ::repairMode 0
  }
  if {![info exists gaSet(Etx220exists)]} {
    set gaSet(Etx220exists) 1
  }